+++ src/core/warnings/readingOrder.ts
@@ -2,6 +2,8 @@ import type { LayoutBlock, PageResult, PageWarning } from '../../types/index.js'
 import { horizontalOverlap } from '../warningTextOverlap.js';
 import { shortTextSample } from './textSamples.js';
 
+export { detectFormLabelReadingOrderDivergence } from './readingOrder/formLabels.js';
+
 /** Reading-order divergence thresholds. A heading the layout pass
  *  places in the first quarter of the page flow but whose text only
  *  shows up in the back half of the native content stream means the
@@ -47,47 +49,6 @@ export function detectReadingOrderDivergence(page: PageResult, blocks: LayoutBlo
   if (blocks.length >= READING_ORDER_MIN_BLOCKS && detectLateBlockStartsNativeText(page, blocks, out)) return;
   if (detectLineReadingOrderDivergence(page, blocks, out)) return;
   detectLocalMathReadingOrderDivergence(page, blocks, out);
-}
-
-export function detectFormLabelReadingOrderDivergence(
-  page: PageResult,
-  blocks: LayoutBlock[],
-  out: PageWarning[],
-): void {
-  if (!page.formFields || page.formFields.length < 2) return;
-  if (out.some((warning) => warning.code === 'reading_order_divergence')) return;
-
-  const labelCounts = new Map<string, number>();
-  for (const field of page.formFields) {
-    const label = collapseReadingOrderWhitespace((field.label?.text ?? '').normalize('NFKC'));
-    if (label.length < 5) continue;
-    labelCounts.set(label, (labelCounts.get(label) ?? 0) + 1);
-  }
-  const uniqueLabels = new Set([...labelCounts].filter(([, count]) => count === 1).map(([label]) => label));
-  if (uniqueLabels.size < 3) return;
-
-  const nativeText = collapseReadingOrderWhitespace(page.text.normalize('NFKC'));
-  const probes = blocks
-    .map((block, index) => ({ block, index, text: collapseReadingOrderWhitespace(block.text.normalize('NFKC')) }))
-    .filter((probe) => uniqueLabels.has(probe.text))
-    .map((probe) => ({ ...probe, nativeIndex: nativeText.indexOf(probe.text) }))
-    .filter((probe) => probe.nativeIndex >= 0);
-  if (probes.length < 3) return;
-
-  let previous = probes[0];
-  for (const probe of probes.slice(1)) {
-    if (probe.nativeIndex + 2 >= previous.nativeIndex) {
-      previous = probe;
-      continue;
-    }
-    out.push({
-      code: 'reading_order_divergence',
-      severity: 'warning',
-      blockIndex: probe.index,
-      message: `form label "${shortTextSample(probe.block.text)}" appears after "${shortTextSample(previous.block.text)}" visually but earlier in the native text stream — native form text order diverges from what a human reads; prefer layout.blocks order when sequence matters`,
-    });
-    return;
-  }
 }
 
 function detectHeadingReadingOrderDivergence(page: PageResult, blocks: LayoutBlock[], out: PageWarning[]): boolean {
